% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{boot_glmnet}
\alias{boot_glmnet}
\title{Compute Bootstrap Confidence Interval and importance measure of glmnet coefficients}
\usage{
boot_glmnet(
  X,
  y,
  B = 500,
  alpha,
  lambda,
  paralel = FALSE,
  workers = 2,
  saveResample = TRUE,
  conf = 0.95,
  ...
)
}
\arguments{
\item{X}{x matrix of predictor variables.See \code{glmnet} for more details.}

\item{y}{y response variable. See \code{glmnet} for more details.}

\item{B}{number of bootstrap replicates. The Default is 500.}

\item{alpha}{The elasticnet's parameter. See \code{glmnet} for more details.}

\item{lambda}{The shrinkage's parameter. See \code{glmnet} for more details.}

\item{paralel}{Use parallel? The default is FALSE.}

\item{workers}{A positive numeric scalar or a function specifying the maximum number of parallel futures that can be active at the same time before blocking. The default is 2.}

\item{saveResample}{Saving Bootstrap resample. The default is TRUE.}

\item{conf}{confidence level with value [0,1]. The default is 0.95.}

\item{...}{Other arguments that can be passed to \code{glmnet}}
}
\value{
an object of class \code{"cv.glmnet.all"} is returned, which is a list
with the ingredients of the cross-validation fit.
\item{model}{a fitted \code{glmnet} object for the full data.}
\item{boot_result}{table contains importance measure, bootstrap Percentile CI, bootstrap bsa CI,bootstrap standard deviation and bootstrap mean }
\item{boot_resample}{All Bootstrap Resampling data}
}
\description{
Compute Bootstrap Confidence Interval and importance measure of glmnet coefficients
}
\examples{
\dontrun{
data("leukemia", package = "varbvs")
X <- leukemia$x
y <- leukemia$y
mod <- cv.glmnet.all(
  x = X, y = y,
  family = "binomial",
  type.measure = "deviance"
)
info_model <- mod$best_param
mod_boot <- boot_glmnet(
  X = X, y = y,
  family = "binomial", alpha = info_model$alpha,
  lambda = info_model$lambda, B = 50
)
}
}
\references{
Hansu Kim, Tae Hee Lee (2021) \emph{A robust elastic net via bootstrap method under sampling uncertainty for significance analysis of high-dimensional design problems, Knowledge-Based Systems,Vol. 225}
}
