% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnet.all.R
\name{cv.glmnet.all}
\alias{cv.glmnet.all}
\title{Does k-fold cross-validation for glmnet, produces a plot, and returns a value for alpha and lambda}
\usage{
cv.glmnet.all(
  alpha = NULL,
  nalpha = 50,
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C"),
  ...
)
}
\arguments{
\item{alpha}{Optional user-supplied a vector of alpha values for which to do k-fold cross-validation; default is NULL, and glmnetExt chooses its own sequence.}

\item{nalpha}{The number of alpha values - default is 50.}

\item{type.measure}{loss to use for cross-validation. see \code{cv.glmnet} for more details.}

\item{\dots}{Other arguments that can be passed to \code{cv.glmnet}}
}
\value{
an object of class \code{"cv.glmnet.all"} is returned, which is a list
with the ingredients of the cross-validation fit.
\item{model}{a fitted \code{glmnet} object for the full data.}
\item{best_param}{value, standard error and confidence interval of alpha and lambda which gives minimum/maximum metric}
\item{all_param}{value, standard error and confidence interval of all alpha and lambda }
\item{type.measure}{loss to use for cross-validation}
}
\description{
Does k-fold cross-validation for glmnet, produces a plot, and returns a value for alpha and lambda
}
\examples{
\dontrun{
data("leukemia",package = "varbvs")
X=leukemia$x
y=leukemia$y
mod <- cv.glmnet.all(x=X,y=y,family="binomial",type.measure = "deviance")
}
}
\seealso{
\code{glmnet::cv.glmnet}
}
