% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boot_glmnet.R
\name{plot.boot_glmnet}
\alias{plot.boot_glmnet}
\title{Plots the bootstrap result.}
\usage{
\method{plot}{boot_glmnet}(x, ...)
}
\arguments{
\item{x}{fitted 'cv.glmnet.all'}

\item{\dots}{Not used. Other arguments to plot.}
}
\description{
Plots the bootstrap result.
}
\examples{
data("leukemia", package = "varbvs")
X <- leukemia$x
y <- leukemia$y
mod <- cv.glmnet.all(x = X, y = y, family = "binomial", type.measure = "deviance")
info_model <- mod$best_param
mod_boot <- boot_glmnet(
  X = X, y = y,
  family = "binomial",
  alpha = info_model$alpha,
  lambda = info_model$lambda, B = 50
)
plot(mod_boot)
}
