% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.glmnet.all.R
\name{plot.cv.glmnet.all}
\alias{plot.cv.glmnet.all}
\title{Plots the cross-validation contour.}
\usage{
\method{plot}{cv.glmnet.all}(
  x,
  color.temp = "temp1",
  n.color.grid = 30,
  reversescale = FALSE,
  label.font.size = 10,
  color.label.font = "white",
  ...
)
}
\arguments{
\item{x}{fitted 'cv.glmnet.all'.}

\item{color.temp}{color template,should be either "temp1","temp2", and "temp3" or vector of colors with length>1. The default is "temp1".}

\item{n.color.grid}{number of color grid. The default is 30.}

\item{reversescale}{should color template reverse order? The default is FALSE}

\item{label.font.size}{font size of label. The default is 10.}

\item{color.label.font}{color of label. The default is "white".}

\item{...}{Not used. Other arguments to plot.}
}
\description{
Plots the cross-validation contour.
}
\details{
The labels are shown on the plot are several non-zero coefficients. Labels are only shown for the top five rankings based on a metric.
}
\examples{
\dontrun{
data("leukemia", package = "varbvs")
X <- leukemia$x
y <- leukemia$y
mod1 <- cv.glmnet.all(x = X, y = y, family = "binomial", type.measure = "deviance")
plot(mod1)
}
}
