% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.glmnet.all.R
\name{predict.cv.glmnet.all}
\alias{predict.cv.glmnet.all}
\title{This function makes it easier to use the results of cross-validation to make a prediction.}
\usage{
\method{predict}{cv.glmnet.all}(object, newx, ...)
}
\arguments{
\item{object}{fitted 'cv.glmnet.all'.}

\item{newx}{A Matrix new values for \code{x} at which predictions are to be.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on the \dots{} argument which is passed.
on to the \code{predict} method for \code{glmnet} objects.
}
\description{
This function makes it easier to use the results of cross-validation to make a prediction.
}
\examples{
data("leukemia", package = "varbvs")
X <- leukemia$x
y <- leukemia$y
mod <- cv.glmnet.all(x = X, y = y, family = "binomial", type.measure = "deviance")
predict(mod, X)
}
